+++ src/main/java/be/ugent/rml/functions/lib/IDLabFunctions.java
@@ -1,9 +1,23 @@
 package be.ugent.rml.functions.lib;
 
+import com.jayway.jsonpath.Configuration;
+import com.jayway.jsonpath.JsonPath;
+import com.jayway.jsonpath.PathNotFoundException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.io.BufferedReader;
+import java.io.InputStreamReader;
+import java.net.HttpURLConnection;
+import java.net.URL;
+import java.net.URLEncoder;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 public class IDLabFunctions {
+
+    private static final Logger logger = LoggerFactory.getLogger(IDLabFunctions.class);
 
     public static boolean stringContainsOtherString(String str, String otherStr, String delimiter) {
         String[] split = str.split(delimiter);
@@ -17,6 +31,55 @@ public class IDLabFunctions {
             return list.contains(str);
         } else {
             return false;
+        }
+    }
+
+    public static List<String> dbpediaSpotlight(String text, String endpoint) {
+        if (!text.equals("")) {
+            try {
+                URL url = new URL(endpoint + "/annotate?text=" + URLEncoder.encode(text, "UTF-8"));
+                HttpURLConnection con = (HttpURLConnection) url.openConnection();
+                con.setRequestMethod("GET");
+                con.setRequestProperty("Accept", "application/json");
+                con.setInstanceFollowRedirects(true);
+
+                BufferedReader in = new BufferedReader(new InputStreamReader(con.getInputStream()));
+                String inputLine;
+                StringBuilder content = new StringBuilder();
+
+                while ((inputLine = in.readLine()) != null) {
+                    content.append(inputLine);
+                }
+
+                in.close();
+                con.disconnect();
+
+                Object document = Configuration.defaultConfiguration().jsonProvider().parse(content.toString());
+                return JsonPath.parse(document).read("$.Resources[*].@URI");
+            } catch (PathNotFoundException e) {
+                // that means no result was found, so that is fine
+                logger.info(e.getMessage(), e);
+            } catch (Exception e) {
+                // that probably means smth is wrong with the DBpedia Spotlight endpoint, so that is fine: log and continue
+                logger.warn(e.getMessage(), e);
+            }
+        }
+
+        return new ArrayList<>();
+    }
+
+    public static Object trueCondition(String bool, String value) {
+        if (bool == null || !bool.equals("true")) {
+            return null;
+        } else {
+            return value;
+        }
+    }
+    public static String decide(String input, String expected, String result) {
+        if (input != null && input.equals(expected)) {
+            return result;
+        } else {
+            return null;
         }
     }
 }